Fix one problem with the schema for SNMPv1 Trap messages.
The Trap itself was missing a `#[rasn(choice)]` on `agent_addr`,
resulting in an unconditional decoding error. Additionally, the schema
for `IpAddress` was given as `[u8; 4]` which rasn derives as `SEQUENCE
(SIZE (4)) OF INTEGER` rather than `OCTET STRING`.

A test is added containing a sample trap packet; this now decodes
without an error but is still not quite correct: the `variable_bindings`
field silently decodes as an empty vector when it should have two
`VarBind`s in it.

The test additionally sketches out an attempt to check round-tripping by
encoding the equivalent struct, but this throws a choice error if there
are any elements in `variable_bindings`. For now, it checks that for the
same Trap without any `VarBind`s, the encoder does not error.