Fix encoding of newtype variants containing choice enums.
The derive macro is extended to allow `#[rasn(choice)]` annotations on
enum variants which are newtypes containing a choice enum. For variants
that are annotated this way, the visitor for those variants emits calls
to `encode` instead of `encode_with_tag`, analogously to how this is
handled for struct fields with a choice enum type.

The failing assertions testing an encode of `Message<Trap>` are
un-commented and now pass.