Fix decoding of newtype variants containing choice enums.
If a choice enum contains a newtype variant that itself contains a
choice enum type, that variant can match an arbitrary number of tags
corresponding to the effective tags of each variant of the _contained_
choice enum type. Because choice enums derive `AsnType::TAG` as `EOC`,
tag-dispatch-based decoding will never dispatch to that variant because
no encoding of any variant of the contained type ever matches, and in
the case of choice enums such as SMIv1 `ObjectSyntax`, where all
variants are newtypes containing choice enums, decode unconditionally
fails with "Invalid CHOICE discriminant".

To fix this, the generated impl of decode is split into two phases:

- First all variants not marked `#[rasn(choice)]` are handled by the
  current strategy of dispatching on tag value.

- Then, all variants that contain a choice enum (marked with
  `#[rasn(choice]`) are handled in the fall-through arm of the match, by
  speculatively trying to decode the contained type of each such
  variant, and decoding as the first variant that succeeds. If none of
  them succeed, only then does the decode fail with "Invalid CHOICE
  discriminant".

In particular, this makes SMIv1 `ObjectSyntax` derive a working `Decode`
impl, which was previously causing SNMPv1 `VarBind` to fail to decode.
The latter is now specifically unit tested, and the last remaining
failing assertion in the test of `Message<Trap>` is un-commented and
passing.