Remove the requirement to tag newtype variants with `#[rasn(choice)]`.
This is done by adding an associated bool const to `AsnType` called
`CHOICE`, which defaults to false but is overridden to true by the
derived impl for choice enums.

Then, instead of chosing the handling of enum variants based on the
attribute at macro-expand time, both choice and non-choice handling code
is inserted into the macro expansion, controlled by conditionals
referencing this const, and the compiler cleans out the dead branches in
DCE.

In particular, for newtype variants, the tag match arm adds a pattern
guard

```
TAG_n if !<ContainedType as rasn::AsnType>::CHOICE => {...}
```

which makes that arm dead code if `ContainedType` is a choice enum.
Then, in the fall-through arm where no tag has matched, a speculative
decode for each newtype variant is inserted, again guarded by a const-if
that makes it dead code if the contained type is not a choice enum:

```
if <ContainedType as rasn::AsnType>::CHOICE {
    if let Ok(val) = ContainedType::decode(decoder) {
        return Ok(Ty::NewtypeVariant(val))
    }
}
```

And finally, for struct variants, each field's decoding contains a
conditional:

```
Ty::StructVariant {
    field: if <FieldType as rasn::AsnType>::CHOICE {
        <_>::decode(...)
    } else {
        <_>::decode_with_tag(...)
    },
}
```