Create TagTree and refactor Tag validation and CHOICE decoding
This changes rasn's Tag validation to use the new `TagTree` data
structure, representing a tree of tags in data structure, this structure
is populated for CHOICE enums to contain their complete variants as well
as structs for checking groups of fields.

Additionally decoding has been changed to use `Tag::EOC` to determine
whether a field or variant is a `choice` or not, which decides how we
decode and validate.

Co-authored-by: Kyle Schaffrick <kyle@raidi.us>